refactor: custom hook 분리, 상수와 타입 별도 파일로 분리, 스크롤 처리
