+++ src/pages/chat-list/constant/chat-list-constant.ts
@@ -1,0 +1,77 @@
+export const FILTERS = ['전체보기', '안 읽음', '예약 확정'];
+
+export const mockup = [
+  {
+    clientId: 1,
+    clientName: '고객이름',
+    tagTitle: '오소리 푸드트럭',
+    lastChat: '혹시 예약 가능할까요?',
+    lastChatTime: '오후 5:40',
+    unreadCount: 12,
+  },
+  {
+    clientId: 2,
+    clientName: '고객이름',
+    tagTitle: '오소리 푸드트럭',
+    lastChat: '혹시 예약 가능할까요?',
+    lastChatTime: '오후 5:40',
+    unreadCount: 122,
+  },
+  {
+    clientId: 3,
+    clientName: '고객이름',
+    tagTitle: '오소리 푸드트럭',
+    lastChat: '혹시 예약 가능할까요?',
+    lastChatTime: '오후 5:40',
+    unreadCount: 0,
+  },
+  {
+    clientId: 4,
+    clientName: '고객이름고객이름고객이름',
+    tagTitle: '오소리 푸드트럭',
+    lastChat:
+      '혹시 예약 가능할까요? 아 안된다고요. 알겠습니다. 몇자 제한입니까',
+    lastChatTime: '오후 5:40',
+    unreadCount: 12,
+  },
+  {
+    clientId: 5,
+    clientName: '고객이름',
+    tagTitle: '오소리 푸드트럭',
+    lastChat: '혹시 예약 가능할까요?',
+    lastChatTime: '오후 5:40',
+    unreadCount: 0,
+  },
+  {
+    clientId: 6,
+    clientName: '고객이름',
+    tagTitle: '오소리 푸드트럭',
+    lastChat: '혹시 예약 가능할까요?',
+    lastChatTime: '오후 5:40',
+    unreadCount: 0,
+  },
+  {
+    clientId: 7,
+    clientName: '고객이름',
+    tagTitle: '오소리 푸드트럭',
+    lastChat: '혹시 예약 가능할까요?',
+    lastChatTime: '오후 5:40',
+    unreadCount: 0,
+  },
+  {
+    clientId: 8,
+    clientName: '고객이름',
+    tagTitle: '오소리 푸드트럭',
+    lastChat: '혹시 예약 가능할까요?',
+    lastChatTime: '오후 5:40',
+    unreadCount: 0,
+  },
+  {
+    clientId: 9,
+    clientName: '고객이름',
+    tagTitle: '오소리 푸드트럭',
+    lastChat: '혹시 예약 가능할까요?',
+    lastChatTime: '오후 5:40',
+    unreadCount: 0,
+  },
+];
