+++ src/pages/chat-list/types/chat-list-type.ts
@@ -1,0 +1,8 @@
+export interface Chat {
+  clientId: number;
+  clientName: string;
+  tagTitle: string;
+  lastChat: string;
+  lastChatTime: string;
+  unreadCount: number;
+};
